Crate rune_core

Minimum support: Rust 1.79+.

Visit the site ğŸŒ — Read the book ğŸ“–

Core components for the Rune Language, an embeddable dynamic programming language for Rust.

Note

YOU ARE NOT supposed to depend on this directly. Doing so might cause dependency errors since its API is not stable.

Re-exports

pub use rune_alloc as alloc;

Structs

FullTypeOf

Full type information.

Hash

The primitive hash that among other things is used to reference items, types, and native functions.

Item

The reference to an ItemBuf.

ItemBuf

The name of an item in the Rune Language.

Params

Helper to register a parameterized function.

Protocol

A built in instance function.

RawStr

A raw static string.

Enums

Component

The component of an item.

ComponentRef

A reference to a component of an item.

Traits

IntoComponent

Trait for encoding the current type into a Component.

IntoHash

Trait for types which can be converted into a Hash.

ToTypeHash

Helper trait used to convert a type into a type hash.