rune_core
Module item

Structs

Item

The reference to an ItemBuf.

ItemBuf

The name of an item in the Rune Language.

Iter

An item over the iterator.

Enums

Component

The component of an item.

ComponentRef

A reference to a component of an item.

Traits

IntoComponent

Trait for encoding the current type into a Component.