rune_core::protocol

Struct Protocol

#[non_exhaustive]
pub struct Protocol {
    pub name: &'static str,
    pub method: Option<&'static str>,
    pub hash: Hash,
    /* private fields */
}

A pre-defined protocol function.

Examples

use rune::runtime::Protocol;

let protocol = Protocol::GET;
assert_eq!(protocol.name, "GET");

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: &'static str

The name of the builtin function.

method: Option<&'static str>

If this protocol defines an associated method, this is the name of that method.

hash: Hash

The hash of the builtin function.

Implementations

impl Protocol

pub const GET: Protocol

The function to access a field.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::GET.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::GET));

pub const SET: Protocol

The function to set a field.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::SET.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::SET));

pub const INDEX_GET: Protocol

The function to access an index.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::INDEX_GET.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::INDEX_GET));

pub const INDEX_SET: Protocol

The function to set an index.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::INDEX_SET.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::INDEX_SET));

pub const PARTIAL_EQ: Protocol

Check two types for partial equality.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::PARTIAL_EQ.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::PARTIAL_EQ));

pub const EQ: Protocol

Check two types for total equality.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::EQ.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::EQ));

pub const PARTIAL_CMP: Protocol

Perform an partial comparison between two values.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::PARTIAL_CMP.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::PARTIAL_CMP));

pub const GT: Protocol

The protocol behind the > operator.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::GT.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::GT));

pub const GE: Protocol

The protocol behind the >= operator.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::GE.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::GE));

pub const LT: Protocol

The protocol behind the > operator.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::LT.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::LT));

pub const LE: Protocol

The protocol behind the <= operator.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::LE.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::LE));

pub const MAX: Protocol

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::MAX.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::MAX));

pub const MIN: Protocol

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::MIN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::MIN));

pub const CMP: Protocol

Perform an total comparison between two values.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::CMP.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::CMP));

pub const ADD: Protocol

The function to implement for the addition operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::ADD.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::ADD));

pub const ADD_ASSIGN: Protocol

The function to implement for the addition assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::ADD_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::ADD_ASSIGN));

pub const SUB: Protocol

The function to implement for the subtraction operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::SUB.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::SUB));

pub const SUB_ASSIGN: Protocol

The function to implement for the subtraction assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::SUB_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::SUB_ASSIGN));

pub const MUL: Protocol

The function to implement for the multiply operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::MUL.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::MUL));

pub const MUL_ASSIGN: Protocol

The function to implement for the multiply assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::MUL_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::MUL_ASSIGN));

pub const DIV: Protocol

The function to implement for the division operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::DIV.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::DIV));

pub const DIV_ASSIGN: Protocol

The function to implement for the division assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::DIV_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::DIV_ASSIGN));

pub const REM: Protocol

The function to implement for the remainder operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::REM.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::REM));

pub const REM_ASSIGN: Protocol

The function to implement for the remainder assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::REM_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::REM_ASSIGN));

pub const BIT_AND: Protocol

The function to implement for the bitwise and operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::BIT_AND.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::BIT_AND));

pub const BIT_AND_ASSIGN: Protocol

The function to implement for the bitwise and assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::BIT_AND_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::BIT_AND_ASSIGN));

pub const BIT_XOR: Protocol

The function to implement for the bitwise xor operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::BIT_XOR.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::BIT_XOR));

pub const BIT_XOR_ASSIGN: Protocol

The function to implement for the bitwise xor assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::BIT_XOR_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::BIT_XOR_ASSIGN));

pub const BIT_OR: Protocol

The function to implement for the bitwise or operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::BIT_OR.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::BIT_OR));

pub const BIT_OR_ASSIGN: Protocol

The function to implement for the bitwise xor assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::BIT_OR_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::BIT_OR_ASSIGN));

pub const SHL: Protocol

The function to implement for the bitwise shift left operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::SHL.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::SHL));

pub const SHL_ASSIGN: Protocol

The function to implement for the bitwise shift left assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::SHL_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::SHL_ASSIGN));

pub const SHR: Protocol

The function to implement for the bitwise shift right operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::SHR.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::SHR));

pub const SHR_ASSIGN: Protocol

The function to implement for the bitwise shift right assign operation.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::SHR_ASSIGN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::SHR_ASSIGN));

pub const DISPLAY_FMT: Protocol

Protocol function used by template strings.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::DISPLAY_FMT.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::DISPLAY_FMT));

pub const DEBUG_FMT: Protocol

Protocol function used by custom debug impls.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::DEBUG_FMT.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::DEBUG_FMT));

pub const INTO_ITER: Protocol

Function used to convert an argument into an iterator.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::INTO_ITER.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::INTO_ITER));

pub const NEXT: Protocol

The function to call to continue iteration.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::NEXT.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::NEXT));

pub const NTH: Protocol

The function to call to continue iteration at the nth element.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::NTH.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::NTH));

pub const NTH_BACK: Protocol

The function to call to continue iteration at the nth element form the back.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::NTH_BACK.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::NTH_BACK));

pub const SIZE_HINT: Protocol

Protocol used when getting the size hint of an iterator.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::SIZE_HINT.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::SIZE_HINT));

pub const LEN: Protocol

Protocol used when getting the exact length of an iterator.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::LEN.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::LEN));

pub const NEXT_BACK: Protocol

Protocol used when cloning a value.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::NEXT_BACK.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::NEXT_BACK));

pub const INTO_FUTURE: Protocol

Function used to convert an argument into a future.

Signature: fn(Value) -> Future.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::INTO_FUTURE.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::INTO_FUTURE));

pub const INTO_TYPE_NAME: Protocol

Coerce a value into a type name. This is stored as a constant.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::INTO_TYPE_NAME.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::INTO_TYPE_NAME));

pub const IS_VARIANT: Protocol

Function used to test if a value is a specific variant.

Signature: fn(self, Hash) -> bool.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::IS_VARIANT.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::IS_VARIANT));

pub const TRY: Protocol

Function used for the question mark operation.

Signature: fn(self) -> Result.

Note that it uses the Result like Try uses ControlFlow i.e., for Result::<T, E> it should return Result<T, Result<(), E>>

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::TRY.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::TRY));

pub const HASH: Protocol

Protocol used when calculating a hash.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::HASH.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::HASH));

pub const CLONE: Protocol

Protocol used when cloning a value.

Examples

use rune::hash::IntoHash;
use rune::runtime::Protocol;

 let hash = Protocol::CLONE.into_hash();
let protocol = Protocol::from_hash(hash);
 assert_eq!(protocol, Some(Protocol::CLONE));

pub const fn from_hash(hash: Hash) -> Option<Self>

Look up protocol for the given hash.

Examples

use rune::runtime::Protocol;

let hash = Protocol::GET.hash;
let protocol = Protocol::from_hash(hash).ok_or("missing protocol")?;
assert_eq!(protocol, Protocol::GET);