Crate rune_core

Minimum support: Rust 1.81+.

Visit the site ğŸŒ — Read the book ğŸ“–

Core components for the Rune Language, an embeddable dynamic programming language for Rust.

Note

YOU ARE NOT supposed to depend on this directly. Doing so might cause dependency errors since its API is not stable.

Re-exports

pub use rune_alloc as alloc;

Modules

hash

item

params

protocol

raw_str