rune_core::hash
Struct Hash

#[repr(transparent)]
pub struct Hash(/* private fields */);

The primitive hash that among other things is used to reference items, types, and native functions.

Implementations

impl Hash

pub const EMPTY: Self

The empty hash.

pub fn as_non_empty(&self) -> Option<Self>

Return the current hash if it is non-empty.

pub fn index(index: usize) -> Self

Construct a hash from an index.

pub fn ident(name: &str) -> Hash

Get the hash corresponding to a string identifier like function or hello_world.

pub fn type_hash(path: impl ToTypeHash) -> Self

Get the hash of a type.

pub fn associated_function( type_hash: impl IntoHash, name: impl IntoHash, ) -> Self

Construct a hash to an instance function, where the instance is a pre-determined type.

pub fn field_function( protocol: impl IntoHash, type_hash: Hash, name: impl IntoHash, ) -> Self

Construct a hash corresponding to a field function.

pub fn index_function( protocol: impl IntoHash, type_hash: Hash, index: Hash, ) -> Self

Construct an index function.

pub fn static_bytes(bytes: &[u8]) -> Hash

Get the hash corresponding to a static byte array.

pub fn object_keys<I>(keys: I) -> Self
where I: IntoIterator, I::Item: AsRef<str>,

Hash the given iterator of object keys.

pub const fn with_generics(self, generics: Self) -> Self

Mix in generics hash.

The generics hash must be a combination of the output from with_type_parameters and with_function_parameters.

pub const fn with_type_parameters(self, ty: Self) -> Self

Mix the current hash with type parameters.

pub const fn with_function_parameters(self, f: Self) -> Self

Mix the current hash with function parameters.

pub const fn parameters<const N: usize>(params: [Hash; N]) -> Self

Hash type parameters.